% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{R6 class representing a simulation of the peppered moth system}
\description{
R6 class representing a simulation of the peppered moth system.
}
\details{
Used to generate a world with N peppered moth individuals that can be either black
or white. Method `run()` can be used to simulate changes in moth colour over time
as the colour of the world changes.
}
\examples{
simu <- simulation$new(years = 200, N = 500, mutation_rate = 1e-2, period = 50)
simu$output
simu$run()
simu$plot()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{world}}{R6 object of class world. World to simulate.}

\item{\code{population}}{R6 object of class population. Population of moths within the world.}

\item{\code{years}}{Integer. Number of years over which to simulate.}

\item{\code{output}}{Data frame. Output of simulation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{simulation$new()}}
\item \href{#method-run}{\code{simulation$run()}}
\item \href{#method-plot}{\code{simulation$plot()}}
\item \href{#method-clone}{\code{simulation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize new R6 object of class 'simulation'.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{simulation$new(years = 10000, N = 100, mutation_rate = 0.001, period = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{years}}{Integer. Number of years over which to simulate.}

\item{\code{N}}{Integer. Size of population within the world.}

\item{\code{mutation_rate}}{Probability. Probability that moths can change colour in a timestep.}

\item{\code{period}}{period Integer. Number of time steps after which colour of world changes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An R6 object of class 'simulation'
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Run simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{simulation$run()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Simulation will run for a number of time steps defined by `years` attribute.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plot simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{simulation$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{simulation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
