% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moth.R
\name{moth}
\alias{moth}
\title{R6 class representing an individual moth}
\description{
R6 class representing an individual moth.
}
\examples{
#Create a world in which the moth lives
example_world <- world$new()

#Create new moth object
example_moth <- moth$new(world = example_world)

#Moth has empty fitness initially
example_moth$fitness

#Compute fitness and then print it
example_moth$computefitness()
example_moth$fitness

## ------------------------------------------------
## Method `moth$mutation`
## ------------------------------------------------

#Moth with low mutation rate is very unlikely to change colour
set.seed(123L)
my_world <- world$new()
low_mutation_moth  <- moth$new(mutation_rate = 1e-10, world = my_world)
low_mutation_moth$colour
low_mutation_moth$mutation()
low_mutation_moth$colour

#Moth with mutation rate 1 is almost guaranteed to change colour
high_mutation_moth <- moth$new(mutation_rate = 1, world = my_world)
high_mutation_moth$colour
high_mutation_moth$mutation()
high_mutation_moth$colour

## ------------------------------------------------
## Method `moth$computefitness`
## ------------------------------------------------

set.seed(123L)
my_world <- world$new()
my_moth  <- moth$new(world = my_world)

#World is black (i.e. colour = 1) and the moth is white (i.e. colour = 0)
#Therefore, fitness is lower (0.5)
my_world$colour
my_moth$colour
my_moth$computefitness()
my_moth$fitness

#If the world were also white, moth fitness is 1.
my_world$colour <- 0
my_moth$computefitness()
my_moth$fitness
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{colour}}{Integer. Colour of the moth. Can be either 0 (white) or 1 (black)}

\item{\code{mutation_rate}}{Probability. Probability of mutation in genes that control colour.}

\item{\code{world}}{R6 object of class world. World in which moth lives.}

\item{\code{fitness}}{Numeric. Fitness of individual in current world. Ranges between 0 and 1.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{moth$new()}}
\item \href{#method-mutation}{\code{moth$mutation()}}
\item \href{#method-computefitness}{\code{moth$computefitness()}}
\item \href{#method-clone}{\code{moth$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize new R6 object of class 'moth'.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{moth$new(mutation_rate = 1e-06, world)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mutation_rate}}{Probability. Probability of mutation in genes that control colour.}

\item{\code{world}}{R6 object of class world. World in which moth lives.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An R6 object of class 'moth'
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mutation"></a>}}
\if{latex}{\out{\hypertarget{method-mutation}{}}}
\subsection{Method \code{mutation()}}{
Change moth colour
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{moth$mutation()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Moth colour will change with a probability defined by attribute
`mutation_rate`
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{#Moth with low mutation rate is very unlikely to change colour
set.seed(123L)
my_world <- world$new()
low_mutation_moth  <- moth$new(mutation_rate = 1e-10, world = my_world)
low_mutation_moth$colour
low_mutation_moth$mutation()
low_mutation_moth$colour

#Moth with mutation rate 1 is almost guaranteed to change colour
high_mutation_moth <- moth$new(mutation_rate = 1, world = my_world)
high_mutation_moth$colour
high_mutation_moth$mutation()
high_mutation_moth$colour
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-computefitness"></a>}}
\if{latex}{\out{\hypertarget{method-computefitness}{}}}
\subsection{Method \code{computefitness()}}{
Compute fitness of individual in current environment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{moth$computefitness()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Compare colour of moth to the colour of the world and update attribute
`fitness` accordingly. Fitness can be either 0.5 or 1.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{set.seed(123L)
my_world <- world$new()
my_moth  <- moth$new(world = my_world)

#World is black (i.e. colour = 1) and the moth is white (i.e. colour = 0)
#Therefore, fitness is lower (0.5)
my_world$colour
my_moth$colour
my_moth$computefitness()
my_moth$fitness

#If the world were also white, moth fitness is 1.
my_world$colour <- 0
my_moth$computefitness()
my_moth$fitness
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{moth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
