% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population}
\alias{population}
\title{R6 class representing population of multiple moths}
\description{
R6 class representing population of multiple moths.
}
\examples{
 set.seed(2L)
 my_world <- world$new()
 my_population <- population$new(N = 4, world = my_world)
 my_population$individuals[[1]]$colour <- 1
 my_population$computefitness()
 print(unlist(lapply(my_population$individuals, function(i) i$colour)))
 my_population$reproduction()
 print(unlist(lapply(my_population$individuals, function(i) i$colour)))

## ------------------------------------------------
## Method `population$computefitness`
## ------------------------------------------------

set.seed(123L)
my_world <- world$new()
my_pop   <- population$new(N = 50, world = my_world)

#Fitness of all moths is NA on initialisation
my_pop$individuals[[1]]$fitness

#Compute fitness of all moths
my_pop$computefitness()

#Plot fitness of moths in the world
all_fitness <- sapply(my_pop$individuals, function(x){x$fitness}) 
hist(all_fitness)

## ------------------------------------------------
## Method `population$mutation`
## ------------------------------------------------

set.seed(123L)
my_world <- world$new()
my_pop   <- population$new(N = 50, mutation_rate = 0.75, world = my_world)

#Colour of moths should be ~equal at start
all_colour <- sapply(my_pop$individuals, function(x){x$colour})
hist(all_colour)

#Change colour of all moths with probability defined by `mutation_rate`
#Colours have changed
my_pop$mutation()
all_colour <- sapply(my_pop$individuals, function(x){x$colour}) 
hist(all_colour)

## ------------------------------------------------
## Method `population$reproduction`
## ------------------------------------------------

set.seed(123L)
my_world <- world$new()
my_pop   <- population$new(N = 50, world = my_world)

#At initialization colour of moths is ~even
all_colour <- sapply(my_pop$individuals, function(x){x$colour})
hist(all_colour)

#The world is black (i.e. colour = 1)
#So black moths will have higher fitness than white moths
my_world$colour
my_pop$computefitness()

#Black moths have higher fitness and more more likely to reproduce
#After reproduction, black individuals are much more common
my_pop$reproduction()
all_colour <- sapply(my_pop$individuals, function(x){x$colour})
hist(all_colour)

## ------------------------------------------------
## Method `population$generation`
## ------------------------------------------------

set.seed(123L)
my_world <- world$new()
my_pop   <- population$new(N = 50, world = my_world)

#At initialization colour of moths is ~even
all_colour <- sapply(my_pop$individuals, function(x){x$colour})
hist(all_colour)

#The world is black (i.e. colour = 1)
#So black moths will have higher fitness than white moths
my_world$colour
my_pop$generation()

#There will be more black moths in the next generation
all_colour <- sapply(my_pop$individuals, function(x){x$colour})
hist(all_colour)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{individuals}}{List. List of R6 objects of class 'moth'.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{population$new()}}
\item \href{#method-computefitness}{\code{population$computefitness()}}
\item \href{#method-mutation}{\code{population$mutation()}}
\item \href{#method-reproduction}{\code{population$reproduction()}}
\item \href{#method-generation}{\code{population$generation()}}
\item \href{#method-clone}{\code{population$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize new R6 object of class 'population'.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$new(N = 100, mutation_rate = 1e-06, world)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{N}}{Integer. Number of individual moths in the population.}

\item{\code{mutation_rate}}{Probability. Probability that moths can change colour in a timestep.}

\item{\code{world}}{R6 object of class world. World in which moths live.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An R6 object of class 'population'
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-computefitness"></a>}}
\if{latex}{\out{\hypertarget{method-computefitness}{}}}
\subsection{Method \code{computefitness()}}{
Compute fitness of all moths in the population.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$computefitness()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Run the method `computefitness` of each moth object in the population.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{set.seed(123L)
my_world <- world$new()
my_pop   <- population$new(N = 50, world = my_world)

#Fitness of all moths is NA on initialisation
my_pop$individuals[[1]]$fitness

#Compute fitness of all moths
my_pop$computefitness()

#Plot fitness of moths in the world
all_fitness <- sapply(my_pop$individuals, function(x){x$fitness}) 
hist(all_fitness)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mutation"></a>}}
\if{latex}{\out{\hypertarget{method-mutation}{}}}
\subsection{Method \code{mutation()}}{
Change colour of all moths.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$mutation()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Run the method `mutation` of each moth object in the population.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{set.seed(123L)
my_world <- world$new()
my_pop   <- population$new(N = 50, mutation_rate = 0.75, world = my_world)

#Colour of moths should be ~equal at start
all_colour <- sapply(my_pop$individuals, function(x){x$colour})
hist(all_colour)

#Change colour of all moths with probability defined by `mutation_rate`
#Colours have changed
my_pop$mutation()
all_colour <- sapply(my_pop$individuals, function(x){x$colour}) 
hist(all_colour)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reproduction"></a>}}
\if{latex}{\out{\hypertarget{method-reproduction}{}}}
\subsection{Method \code{reproduction()}}{
Moths reproduce with a probability defined by their fitness in the environment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$reproduction()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Offspring are clones of their parents (i.e. they have the same colour!)
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{set.seed(123L)
my_world <- world$new()
my_pop   <- population$new(N = 50, world = my_world)

#At initialization colour of moths is ~even
all_colour <- sapply(my_pop$individuals, function(x){x$colour})
hist(all_colour)

#The world is black (i.e. colour = 1)
#So black moths will have higher fitness than white moths
my_world$colour
my_pop$computefitness()

#Black moths have higher fitness and more more likely to reproduce
#After reproduction, black individuals are much more common
my_pop$reproduction()
all_colour <- sapply(my_pop$individuals, function(x){x$colour})
hist(all_colour)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generation"></a>}}
\if{latex}{\out{\hypertarget{method-generation}{}}}
\subsection{Method \code{generation()}}{
Run all demographic methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$generation()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Compute fitness, mutation and reproduction for all moths.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{set.seed(123L)
my_world <- world$new()
my_pop   <- population$new(N = 50, world = my_world)

#At initialization colour of moths is ~even
all_colour <- sapply(my_pop$individuals, function(x){x$colour})
hist(all_colour)

#The world is black (i.e. colour = 1)
#So black moths will have higher fitness than white moths
my_world$colour
my_pop$generation()

#There will be more black moths in the next generation
all_colour <- sapply(my_pop$individuals, function(x){x$colour})
hist(all_colour)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
